websocket_service/requirements.txt
# Core WebSocket Framework
fastapi==0.115.5
websockets==14.1
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Video Processing
av==13.1.0  # PyAV for H.264 encoding

# Async Support
aiofiles==24.1.0

# Utilities
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Type hints
typing-extensions==4.12.2